Fix an issue with identical URLs for GET and POST
If you have two endpoints with identical URLs, but one was a GET
and one was a POST, you would get a 500 error when doing a
GET against the URL.

The problem was passing `request.post` to `normalizeEOL`, which
expects a string. When doing a GET, `request.post` is null,
causing an exception.